Cleaned up warnings, FINALLY disabled warning 4005.

+++ COMP3501/COMP3501/graphics.cpp
@@ -449,7 +449,7 @@ bool Graphics::Frame(int fps, int cpu, float time, Input* input) {
 		m_Tank->shoot();	
 	}
 
-	m_Tank->orientTurret(deltaX,deltaY,time);
+	m_Tank->orientTurret((float)deltaX, (float)deltaY, time);
 
 	m_Tank->Update(input, time, m_QuadTree);
 	for (int i=0; i<NUM_ENEMYS; i++)
